Add AI validation workflow and stabilize CI audit
