+++ .github/workflows/supabase-ci-cd.yml
@@ -43,8 +43,8 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Audit dependencies
-        run: npm audit --audit-level=moderate
+      - name: Audit dependencies (high+)
+        run: npm audit --audit-level=high
 
       - name: Run unit tests
         run: npm test -- --runInBand
